Change some naming that wasn't super intellegent
